scrapping/requirements.txt
requests_html
lxml_html_clean
pymysql
python-dotenv
requests
pandas
selenium
webdriver-manager
chromedriver-autoinstaller
pyarrow
redis